auth added

+++ routes/cartRouter.js
@@ -5,32 +5,74 @@ const cartRouter = new Router();
 
 cartRouter.post("/create", async(req,res) => {
     try {
-        const response = await create(req);
-        res.send(response)
+        if(req.isAuth) {
+            if(req.access === "customer") {
+                const response = await create(req);
+                res.send(response)
+            } else {
+                const err = new Error("Unauthorized Access");
+                res.send({ err: err.message });
+                res.end();
+            }
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err: error.message})
     }
 })
 cartRouter.patch("/update/:cartId", async(req,res) => {
     try {
-        const response = await update(req);
-        res.send(response)
+        if(req.isAuth) {
+            if(req.access === "customer") {
+                const response = await update(req);
+                res.send(response)
+            } else {
+                const err = new Error("Unauthorized Access");
+                res.send({ err: err.message });
+                res.end();
+            }
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err: error.message})
     }
 })
 cartRouter.delete("/delete/:cartId", async (req,res) => {
     try {
-        const response = await deleteCart(req);
-        res.send(response)
+        if(req.isAuth) {
+            if(req.access === "customer") {
+                const response = await deleteCart(req);
+                res.send(response)
+            } else {
+                const err = new Error("Unauthorized Access");
+                res.send({ err: err.message });
+                res.end();
+            }
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err: error.message})
     }
 })
 cartRouter.get("/:cartId", async(req,res) => {
     try {
-        const response = await get(req);
-        res.send(response)
+        if(req.isAuth) {
+            const response = await get(req);
+            res.send(response)
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err: error.message})
     }
