+++ routes/couponRouter.js
@@ -5,9 +5,21 @@ const couponRouter = new Router();
 
 couponRouter.post("/add",async(req,res)=> {
     try {
-        const response = await addCoupon(req);
-        res.send(response);
-        res.end();
+        if(req.isAuth) {
+            if(req.access === "admin") {
+                const response = await addCoupon(req);
+                res.send(response);
+                res.end();
+            } else {
+                const err = new Error("Unauthorized Access");
+                res.send({ err: err.message });
+                res.end();
+            }
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err:error.message});
         res.end();
@@ -15,9 +27,15 @@ couponRouter.post("/add",async(req,res)=> {
 })
 couponRouter.get("/:couponId",async(req,res)=> {
     try {
-        const response = await get(req);
-        res.send(response);
-        res.end();
+        if(req.isAuth) {
+                const response = await get(req);
+                res.send(response);
+                res.end();
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err:error.message});
         res.end();
@@ -25,9 +43,21 @@ couponRouter.get("/:couponId",async(req,res)=> {
 })
 couponRouter.put("/update/:couponId",async(req,res)=> {
     try {
-        const response = await updateCoupon(req);
-        res.send(response);
-        res.end();
+        if(req.isAuth) {
+            if(req.access === "admin") {
+                const response = await updateCoupon(req);
+                res.send(response);
+                res.end();
+            } else {
+                const err = new Error("Unauthorized Access");
+                res.send({ err: err.message });
+                res.end();
+            }
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err:error.message});
         res.end();
@@ -35,9 +65,21 @@ couponRouter.put("/update/:couponId",async(req,res)=> {
 })
 couponRouter.delete("/delete/:couponId",async(req,res)=> {
     try {
-        const response = await deleteCoupon(req);
-        res.send(response);
-        res.end();
+        if(req.isAuth) {
+            if(req.access === "admin") {
+                const response = await deleteCoupon(req);
+                res.send(response);
+                res.end();
+            } else {
+                const err = new Error("Unauthorized Access");
+                res.send({ err: err.message });
+                res.end();
+            }
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
         res.send({err:error.message});
         res.end();
@@ -45,12 +87,18 @@ couponRouter.delete("/delete/:couponId",async(req,res)=> {
 })
 couponRouter.get("/all",async(req,res)=> {
     try {
-        const response = await getAll(req);
-        res.send(response);
-        res.end();
+        if(req.isAuth) {
+            const response = await getAll(req);
+            res.send(response);
+            res.end();
+        } else {
+            const err = new Error("Unauthenticated Access");
+            res.send({ err: err.message });
+            res.end();
+        }
     } catch (error) {
-        res.send({err:error.message});
         res.end();
+        res.send({err:error.message});
     }
 })
